% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepsubs.R
\name{make.var}
\alias{make.var}
\title{Make tryptic peptide variants}
\usage{
make.var(sequence, cutoff = 4, export = TRUE, filename = "default")
}
\arguments{
\item{sequence}{Character string containig the protein sequence in one letter format}

\item{cutoff}{Variant peptides shorter than the value stored in cutoff are excluded from the list. Default is 4. Wild-type peptides are not excluded irrespective of length.}

\item{filename}{Output filename to be used if save = TRUE. By default,  this will be the date followed by the name of the analysed variable and '.Rds'.}

\item{save}{Boolean defaulting to TRUE. If TRUE, the list structure containing the output file will be written to file.}
}
\value{
a list containing a dataframe with the tryptic peptide sequence variants as the first list element.
}
\description{
\code{make.var} generates a list of tryptic peptides and all their possible single amino acid variants from an input protein sequence
}
\examples{

make.var(seq)

make.var(seq, cutoff = 2, filename = 'test.rds')

makevar(seq, export = FALSE)

}

