% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepsubs.R
\name{makevar}
\alias{makevar}
\title{Make tryptic peptide variants}
\usage{
makevar(sequence, cutoff = 4, export = TRUE, filename = "default")
}
\arguments{
\item{sequence}{Character string containig the protein sequence in one letter format}

\item{cutoff}{Peptides shorter than the value stored in cutoff are excluded from the list. Default is 4.}

\item{filename}{Output filename to be used if save = TRUE. By default,  this will be the date followed by the name of the analysed variable and '.Rds'.}

\item{save}{Boolean defaulting to TRUE. If TRUE, the list structure containing the output file will be written to file.}
}
\value{
a list containing a dataframe with the tryptic peptide sequence variants as the first list element.
}
\description{
\code{make.var} generates a list of tryptic peptides and all their possible single amino acid variants from an input protein sequence
}

