% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepanal.R
\name{match.subs}
\alias{match.subs}
\title{Map sites of observed substitutions from list of detected peptides.}
\usage{
match.subs(filename, save = TRUE)
}
\arguments{
\item{filename}{The name of the RDS file which holds the petide and MS data.}

\item{save}{Indicates whether to save the RDS file with the added data back to disk (default = TRUE).}
}
\value{

}
\description{
\code{match.subs} takes the list of observed peptides and maps sites of substitutions to the protein sequence.
}

