% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepclean.R
\name{MS.condense}
\alias{MS.condense}
\title{Sum spectral counts for different ions of the same peptide.}
\usage{
MS.condense(MSdf, repeats = 3, averagerepeats = TRUE, samplenames = "")
}
\arguments{
\item{MSdf}{A dataframe containing MS data (One column for Sequence, one or more columns for spectral counts).}

\item{repeats}{The number of repeat measurements for each sample.}

\item{averagerepeats}{Determines whether repeats should be averaged, default is TRUE.}

\item{samplenames}{A vector of names to be used for the samples.}
}
\value{

}
\description{
\code{MS.condense} sums spectral counts for differently charged ions for the same peptide.
}

