% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepclean.R
\name{MS.load}
\alias{MS.load}
\title{Read in MS results from files or data frames}
\usage{
MS.load(MSfilename, varfilename, scorecutoff = 4.9, scorecol = 8,
  sequencecol = 9, datarange = 11:28, saveresult = TRUE,
  outfilename = "default")
}
\arguments{
\item{MSfilename}{The name of the dataframe or file containing the MS data.}

\item{varfilename}{The name of the file containing the variant peptide data (output of make.var())}

\item{scorecutoff}{The MASCOT score value below which detected peptides are excluded from the analysis (4.9 by default).}

\item{scorecol}{The number of the column containg the MASCOT scores.}

\item{sequencecol}{The number of the column containing the sequences of the detected peptides.}

\item{datarange}{The numbers of the columns containing the spectral counts for the detected peptides (one column per sample)}

\item{saveresult}{Indicates whether to output the result to file, defaults to TRUE}

\item{outfilename}{The name of the file that will be generated if saveresult is TRUE. The filename will be constructed based on varfilename if not specified.}
}
\value{

}
\description{
\code{MS.load} reads in peptide information from a dataframe located in the workspace, or from a .csv file located in the working directory.
}

