% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepclean.R
\name{MS.read}
\alias{MS.read}
\title{Read in MS results from files or data frames}
\usage{
MS.read(MSfilename, scorecutoff = 4.9, scorecol = 8, sequencecol = 9,
  datarange = 11:28)
}
\arguments{
\item{MSfilename}{The name of the dataframe or file containing the MS data.}

\item{scorecutoff}{The MASCOT score value below which detected peptides are excluded from the analysis (4.9 if not specified).}

\item{scorecol}{The number of the column containg the MASCOT scores.}

\item{sequencecol}{The number of the column containing the sequences of the detected peptides.}

\item{datarange}{The numbers of the columns containing the spectral counts for the detected peptides (one column per sample)}
}
\value{

}
\description{
\code{MS.read} reads in peptide information from a dataframe located in the workspace, or from a .csv file located in the working directory.
}
