% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepclean.R
\name{MS.save}
\alias{MS.save}
\title{Add the peptide data to the results list.}
\usage{
MS.save(MSdata, filename)
}
\arguments{
\item{MSdata}{The workpace variable containing the clean MS data.}

\item{filename}{The name of the file to which the cleaned-up MS data will be appended.}
}
\value{

}
\description{
\code{MS.save} loads a peptide file, appends the MS data as a new list item, and saves the file back to its original filename.
}

