% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepplot.R
\name{plot.counts}
\alias{plot.counts}
\title{Plot the relative spectral counts for sustitute peptides}
\usage{
\method{plot}{counts}(filename, excludetop = 0)
}
\arguments{
\item{filename}{The name of the file that contains the Mascot peptide sequences and substitution data.}

\item{excludetop}{The number of high-spectral-count peptides to exclude, counted from the top}
}
\value{

}
\description{
\code{plot.counts} Plots cumulative bar graphs for spectral counts of substituted peptides, normalised to the corresponding wild type peptide.
}

