% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepclean.R
\name{readMS}
\alias{readMS}
\title{Read in CSV files containing MS results}
\usage{
readMS(MSfilename, scorecutoff = 4.9, scorecol = 9, sequencecol = 10,
  datarange = 20:37)
}
\arguments{
\item{MSfilename}{The name of the file containing the MS data.}

\item{scorecutoff}{The score value below which detected peptides are excluded from the analysis.}

\item{scorecol}{The number of the column containg the MASCOT scores.}

\item{sequencecol}{The number of the column containing the sequences of the detected peptides.}

\item{datarange}{The numbers of the columns containing the spectral counts for the detected peptides.}

\item{pepfilename}{The name of the .Rds file containing the peptide list with which the MS data were generated. MS data are added as a new list item.}
}
\value{

}
\description{
\code{readMS} reads in CSV files with lists of detected peptide variants (typically prepared using the Mass Spec-associated software).
}

