% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepsubs.R
\name{save.varcsv}
\alias{save.varcsv}
\title{Export tryptic peptide variants to fasta file}
\usage{
save.varcsv(infilename, outfilename = "default")
}
\arguments{
\item{infilename}{A file containing the output from the make.var function.}

\item{outfilename}{The filename of the csv file. If unspecified, a filename will be constructed based on infilename by replacing the ".Rds" suffix with ".csv".}
}
\value{
The function outputs a csv file containing the peptide sequences and details (site and nature of the substitution) from the specified infile.
}
\description{
\code{save.varcsv} writes the peptides in a variant list into a fasta file
}

