% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepsubs.R
\name{save.varfasta}
\alias{save.varfasta}
\title{Export tryptic peptide variants to fasta file}
\usage{
save.varfasta(infilename, outfilename = "default")
}
\arguments{
\item{sourcevars}{A list output by the trypvar function.}

\item{filename}{The filename of the fasta file. If unspecified, a filename will be conructed based on the input variable name.}
}
\value{
None. The function outputs a fasta-formatted file containing the peptide sequences from the specified variable.
}
\description{
\code{save.varfasta} writes the petides in a variant list into a fasta file
}

