% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepplot.R
\name{plot.cover}
\alias{plot.cover}
\title{Plot the coverage in terms of wild-type peptides}
\usage{
\method{plot}{cover}(filename)
}
\arguments{
\item{filename}{The name of the file that contains the Mascot peptide sequences and substitution data.}
}
\value{

}
\description{
\code{plot.cover} Plots the coverage of the protein sequence in terms of observed wild-type peptides, as well as locations of observed substitutions.
}

