% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepsubs.R
\name{save.varfasta}
\alias{save.varfasta}
\title{Export tryptic peptide variants to fasta file}
\usage{
save.varfasta(infilename, outfilename = "default")
}
\arguments{
\item{infilename}{A file containing the output from the make.var function.}

\item{outfilename}{The filename of the fasta file. If unspecified, a filename will be constructed based on infilename by replacing the ".Rds" suffix with ".fasta".}
}
\value{
The function outputs a fasta-formatted file containing the peptide sequences from the specified infile.
}
\description{
\code{save.varfasta} writes the petides in a variant list into a fasta file
}
